shinyHeatmaply – a shiny app for creating interactive cluster heatmaps
https://www.r-bloggers.com/shinyheatmaply-a-shiny-app-for-creating-interactive-cluster-heatmaps/

The app introduces a functionality that saves to disk a self contained
copy of the htmlwidget as an html file with your data and specifications
you set from the UI, so it can be embedded in webpages, blogposts and
online web appendices for academic publications.

The application has an import interface as part of the application which
currently supports csv, txt, tab, xls, xlsx, rd, rda.